backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
psycopg2-binary==2.9.10
pyyaml==6.0.2
PyJWT==2.10.1
opencv-python-headless==4.12.0.88
numpy==2.2.6
httpx==0.28.1
onnxruntime==1.23.2
